Refactor session layers and move token signer
