+++ api/src/main/java/org/heureum/api/session/api/dto/ActiveSessionsResponse.java
@@ -1,4 +1,4 @@
-package org.heureum.api.stream;
+package org.heureum.api.session.api.dto;
 
 public record ActiveSessionsResponse(String streamId, int activeSessions) {
 }
