+++ api/src/main/java/org/heureum/api/session/infra/jpa/package-info.java
@@ -1,0 +1,3 @@
+package org.heureum.api.session.infra.jpa;
+
+// JPA adapters for session persistence will live here.
