+++ common/src/main/java/org/heureum/common/security/PlaylistTokenSigner.java
@@ -1,4 +1,4 @@
-package org.heureum.stream.security;
+package org.heureum.common.security;
 
 import javax.crypto.Mac;
 import javax.crypto.spec.SecretKeySpec;
