+++ common/src/test/java/org/heureum/common/security/PlaylistTokenSignerTest.java
@@ -1,4 +1,4 @@
-package org.heureum.stream.security;
+package org.heureum.common.security;
 
 import org.junit.jupiter.api.Test;
