+++ core/src/main/java/org/heureum/core/domain/session/StreamSessionStatus.java
@@ -1,4 +1,4 @@
-package org.heureum.core.stream;
+package org.heureum.core.domain.session;
 
 public enum StreamSessionStatus {
     ACTIVE,
