+++ core/src/main/java/org/heureum/core/repository/session/StreamSessionRepository.java
@@ -1,4 +1,6 @@
-package org.heureum.core.stream;
+package org.heureum.core.repository.session;
+
+import org.heureum.core.domain.session.StreamSession;
 
 import java.util.List;
 import java.util.Optional;
