+++ core/src/test/java/org/heureum/core/service/session/StreamSessionServiceTest.java
@@ -1,10 +1,17 @@
-package org.heureum.core.stream;
+package org.heureum.core.service.session;
 
 import org.heureum.common.time.TimeProvider;
+import org.heureum.core.domain.session.StreamSession;
+import org.heureum.core.domain.session.StreamSessionStatus;
+import org.heureum.core.repository.session.StreamSessionRepository;
 import org.junit.jupiter.api.Test;
 
 import java.time.Instant;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
+import java.util.concurrent.ConcurrentHashMap;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -89,5 +96,31 @@ class StreamSessionServiceTest {
                 () -> service.heartbeat(UUID.randomUUID())
         );
         assertNotNull(notFound.getMessage());
+    }
+
+    private static class InMemoryStreamSessionRepository implements StreamSessionRepository {
+        private final ConcurrentHashMap<UUID, StreamSession> sessions = new ConcurrentHashMap<>();
+
+        @Override
+        public StreamSession save(StreamSession session) {
+            sessions.put(session.getSessionId(), session);
+            return session;
+        }
+
+        @Override
+        public Optional<StreamSession> findById(UUID sessionId) {
+            return Optional.ofNullable(sessions.get(sessionId));
+        }
+
+        @Override
+        public List<StreamSession> findByStreamId(String streamId) {
+            List<StreamSession> result = new ArrayList<>();
+            for (StreamSession session : sessions.values()) {
+                if (session.getStreamId().equals(streamId)) {
+                    result.add(session);
+                }
+            }
+            return result;
+        }
     }
 }
